Show correct snackbar message when offline nodes are saved successfully to device

+++ app/src/main/java/mega/privacy/android/app/components/saver/OfflineNodeSaver.kt
@@ -7,6 +7,7 @@ import mega.privacy.android.app.DatabaseHandler
 import mega.privacy.android.app.MegaOffline
 import mega.privacy.android.app.R
 import mega.privacy.android.app.utils.FileUtil.copyFile
+import mega.privacy.android.app.utils.FileUtil.fileExistsInTargetPath
 import mega.privacy.android.app.utils.LogUtil.logError
 import mega.privacy.android.app.utils.OfflineUtils.getOfflineFile
 import mega.privacy.android.app.utils.OfflineUtils.getTotalSize
@@ -27,7 +28,11 @@ class OfflineNodeSaver @Inject constructor(
      * @param highPriority whether this download is high priority or not
      * @param activityStarter a high-order function to launch activity when needed
      */
-    fun save(nodes: List<MegaOffline>, highPriority: Boolean, activityStarter: (Intent, Int) -> Unit) {
+    fun save(
+        nodes: List<MegaOffline>,
+        highPriority: Boolean,
+        activityStarter: (Intent, Int) -> Unit
+    ) {
         save(activityStarter) {
             var totalSize = 0L
             for (node in nodes) {
@@ -42,17 +47,55 @@ class OfflineNodeSaver @Inject constructor(
         externalSDCard: Boolean,
         sdCardOperator: SDCardOperator?
     ) {
-        for (node in (saving as OfflineSaving).nodes) {
-            doDownload(
+        var totalFiles = 0
+        var alreadyExists = 0
+
+        val nodes = (saving as OfflineSaving).nodes
+        for (node in nodes) {
+            val res = doDownload(
                 getOfflineFile(context, node),
                 parentPath, externalSDCard, sdCardOperator
             )
+
+            totalFiles += res.first
+            alreadyExists += res.second
         }
 
         runOnUiThread {
-            showSnackbar(
-                context, context.resources.getQuantityString(R.plurals.download_finish, 1, 1)
-            )
+            if (totalFiles == 0) {
+                showSnackbar(
+                    context,
+                    context.resources.getQuantityString(
+                        R.plurals.empty_folders, nodes.size
+                    )
+                )
+            } else if (totalFiles == alreadyExists) {
+                if (totalFiles == 1) {
+                    showSnackbar(
+                        context, context.resources.getString(R.string.general_already_downloaded)
+                    )
+                } else {
+                    showSnackbar(
+                        context,
+                        context.resources.getQuantityString(
+                            R.plurals.file_already_downloaded, totalFiles, totalFiles
+                        )
+                    )
+                }
+            } else {
+                if (totalFiles == 1) {
+                    showSnackbar(
+                        context, context.resources.getString(R.string.copy_already_downloaded)
+                    )
+                } else {
+                    showSnackbar(
+                        context,
+                        context.resources.getQuantityString(
+                            R.plurals.download_finish, totalFiles, totalFiles
+                        )
+                    )
+                }
+            }
         }
     }
 
@@ -61,7 +104,7 @@ class OfflineNodeSaver @Inject constructor(
         parentPath: String,
         externalSDCard: Boolean,
         sdCardOperator: SDCardOperator?
-    ) {
+    ): Pair<Int, Int> {
         if (file.isDirectory) {
             val dstDir = File(parentPath, file.name)
 
@@ -69,20 +112,39 @@ class OfflineNodeSaver @Inject constructor(
                 dstDir.mkdirs()
             }
 
-            val children = file.listFiles() ?: return
+            val children = file.listFiles() ?: return Pair(0, 0)
+
+            var totalFiles = 0
+            var alreadyExists = 0
 
             children.forEach {
-                doDownload(it, dstDir.absolutePath, externalSDCard, sdCardOperator)
+                val res = doDownload(it, dstDir.absolutePath, externalSDCard, sdCardOperator)
+                totalFiles += res.first
+                alreadyExists += res.second
             }
+
+            return Pair(totalFiles, alreadyExists)
         } else {
             if (externalSDCard && sdCardOperator != null) {
+                if (sdCardOperator.fileExistsInTargetPath(file, parentPath)) {
+                    return Pair(1, 1)
+                }
+
                 try {
                     sdCardOperator.moveFile(parentPath, file)
                 } catch (e: Exception) {
                     logError("Error moving file to the sd card path with exception", e);
                 }
+
+                return Pair(1, 0)
             } else {
+                if (fileExistsInTargetPath(file, parentPath)) {
+                    return Pair(1, 1)
+                }
+
                 copyFile(file, File(parentPath, file.name))
+
+                return Pair(1, 0)
             }
         }
     }
